Move VisitorFamily into extras.
