+++ extras/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 module.exports = {
-  toAST: require('./semantics-toAST').helper,
-  semanticsForToAST: require('./semantics-toAST').semantics
+  VisitorFamily: require('./VisitorFamily'),
+  semanticsForToAST: require('./semantics-toAST').semantics,
+  toAST: require('./semantics-toAST').helper
 };
